Move ItemPagingSource to LazyList and attribute LazyListIntervalContent to Android
