+++ redwood-lazylayout-compose/build.gradle
@@ -30,6 +30,7 @@ spotless {
   kotlin {
     targetExclude(
       // Apache 2-licensed files from AOSP.
+      "src/commonMain/kotlin/app/cash/redwood/lazylayout/compose/LazyListIntervalContent.kt",
       "src/commonMain/kotlin/app/cash/redwood/lazylayout/compose/layout/IntervalList.kt",
       "src/commonMain/kotlin/app/cash/redwood/lazylayout/compose/layout/LazyLayoutIntervalContent.kt",
     )
